Polymorphic Identifiers -- *migration required*
First commit to add polymorphic identifier functionality to all model objects.  Here, I'm adding the database tables, the models, an association to Identifiers in the Publications model, and a rake task to demonstrate functionality.  Currently, we capture raw identifier strings in the Publication.issn_isbn field.  This code and rake task will evaluate the raw issn_isbn field, validate any isbns or issns discovered and add the Identifiers to the Publication model.  

1) Run the database migration via: rake db:migrate
2) Run the rake task: rake identifiers:validate_issn_isbns