Home Pages -- *solr restart required*
Added two new fields to the solr mappings from the Works table:
	created_at
	updated_at

Both are date/time values, and are useful for ordering the Recent 
Additions and Recent Works list on the home pages.

Also, changed the way the faces are shown/organized on the home
pages. Now doing it the Rails Way (not from solr), so even people 
with no works will show up in the list. These too are orderded by 
most recent work entered into BibApp. 

git-svn-id: https://bibapp.googlecode.com/svn/trunk@1114 6a0f952e-c525-0410-88e5-87db5789ed7d