Publisher & Publication Authorities: *==== Server Restart Required ====*
Updated the Admin Publisher Authorities and Admin Publication Authorities
views so that they look the same (i.e., same columns, order) as the Publisher
and Publication index pages. 

Also corrected the Add to Box / Remove from Box functions, removing some
buggy Javascript. Now the page refreshes every time a record is added or
removed from the authority box. It's not as elegant, but it works consistently.