Bug fix (Issue 71): Marshal data too short
I was using session variables to hold the contents of the 
Authority Records Box for Publishers and Publications. It was 
reported that the data was in some cases too big for the session 
variable. 

I changed the code so that only record IDs are stored in session vars.
Hopefully this will clear up the problem.

NB: Before you update your BibApp with this revision, make sure the 
Authority Records Box is empty for both Publishers and Publications
(i.e., the session variables are clear) at:

/publishers/authorities
/publications/authorities

git-svn-id: https://bibapp.googlecode.com/svn/trunk@1139 6a0f952e-c525-0410-88e5-87db5789ed7d