Pen names: Add a name string
Adds the name string to the system based on machine
name, but then sets the name exactly as was typed here.

Before, we would create the name string with machine
name, and then the name was created automatically from
parsing the machine name. This is trouble for people with
two-word last names.

Ex: La Barre, Kathryn was getting parsed as La, Barre Kathryn