Use Rails 3.0.4. This seems to solve the problem I was having with accessing anything through the server (which was giving a no method error on NilClass for 'keys').
The tracing I did of the error led deep into the internals of Rails, so
I decided to try a later version to see if it was a known problem that
was fixed, and since the site appears to work now, I conclude that it
probably was.

I think now it will be possible to get on to fixing application code
and deprecations - the root page already shows up some of the needed
changes.