Revert back person.rb.
There was a deeper problem in machine_name. When creating a machine_name
at some point the method started returning an ActiveSupport::MultiByte::Char
object instead of an actual String. For most purposes this was fine, but
when doing a uniq on a collection, even when the wrapped strings were the
same the ActiveSupport::MultiByte::Chars objects are different, so
the collection stopped compressing as it did before.
Fixed by converting the return of make_machine_name back to an actual
String.