For the contributorships/admin view (unverified contributorships) made the table into a Javascript table. This is mainly as a test case for doing this since this table is simple. Things involved: - Added Fugue icon set - Added DataTables js library - Altered HTML for the table to conform to Datatable's requirements - Added necessary javascript and css includes to view - Added some minimal css to style table - Added javascript to initialize table
Next will be to play some more with the css. I did it all from scratch and
it may be better to look at the sample css file that came with DataTables
and see what can be done with that.

It does work nicely and almost trivially (once you figure it out) allows for
easily getting sorting, pagination, and searching of existing tables.

If we don't have the full table it should be fairly easy to implement the
fetching from the server for pieces, but that's not important with this
particular table, so it can wait for later. But it's in there.