Specify ruby in .rvmrc. Remove libxml2-ruby from Gemfile (let solr-ruby force its inclusion). Remove solr-ruby version.
At this point, solr-ruby does try to include libxml-ruby, but version 0.8.3
which is very old and ruby 1.9.3 doesn't want to build.

So I'm going to drop back to master and see if solr-ruby, which is dead,
can be replaced by another low level library, e.g. rsolr (which would be nice
as that's what sunspot uses).