Checkpointing - CSV load will work now if we remove the delay, but fails with the delay in place.
The delayed job gets enqueued and a worker can take it off and 'run'
it, but as far as I can tell nothing actually happens when the
'running' occurs. The file system side effects in this commit
to test that don't happen. On the other hand, no error is actually
recorded - as far as delayed job is concerned the job actually seems
to run without incident.

Two thoughts:
- it may be that things would work if the CsvPeopleUpload class were
an AR subclass
- it may be that there is a problem finding this stuff in the lib
directory when delayed job tries to run - the includes may be off.